requirements.txt
python-dotenv==1.0.0
openai==1.3.0
pandas==2.1.0
numpy==1.24.3
chromadb==0.4.22
langchain==0.1.0
langchain-openai==0.0.2
jupyter==1.0.0
ipywidgets==8.1.0
scikit-learn==1.3.0
scipy==1.11.0
matplotlib==3.8.0
seaborn==0.13.0
tqdm==4.66.0
pytest==7.4.0
